moving fix/assert functions to their own namespace, "ScalarTransport"
